requirements.txt
streamlit 
numpy
pandas
streamlit-aggrid
matplotlib
seaborn 
streamlit-option-menu
streamlit-tags
pillow
plotly
pycountry
sklearn
scipy
openpyxl
streamlit-drawable-canvas
